make it possible for vitest tests that do not need to start services
Most of the other components take pieces of hubstate and could be tested with a simple mock of hubMessages.ts.  Moves the startup of services to a test helper that can be optionally run from a vitest test modules beforeAll.
