+++ src/webapp_vite/testHelpers/globalContext.ts
